refactor: standardize MissingAttributeError to cclib namespace
Define MissingAttributeError once in cclib/__init__.py and import it
in all files that previously defined their own duplicate class.

Removed duplicate class definitions from:
  - cclib/bridge/cclib2pyquante.py
  - cclib/bridge/cclib2chemfiles.py
  - cclib/bridge/cclib2pyscf.py
  - cclib/method/calculationmethod.py
  - cclib/io/filewriter.py
  - cclib/io/wfxwriter.py
  - cclib/method/population.py

Updated test imports to use the central location:
  - test/method/testbader.py
  - test/method/testhirshfeld.py
  - test/method/testddec.py
  - test/method/testpopulation.py
  - test/io/testwfxwriter.py
  - test/io/testmoldenwriter.py

Fixes #1746
