+++ cclib/bridge/cclib2pyscf.py
@@ -18,10 +18,7 @@ import periodictable
 
 l_sym2num = {"S": 0, "P": 1, "D": 2, "F": 3, "G": 4}
 
-
-class MissingAttributeError(Exception):
-    pass
-
+from cclib import MissingAttributeError
 
 _found_pyscf = find_package("pyscf")
 if _found_pyscf:
